refactor forms
